Begin conversion to Typescript
The app builds in both prod and dev modes, however there are still a few nice things that have not been ported over to Typescriptland.

Especially:

- Hot module replacement
- Redux dev tools

Also, there are several spots where type is set to `any`. This is not good, obvs, but it's meant as a placeholder while we get this things wired up. These `any`s need to be swapped out with specific types. Search for `TODO` to get a list of these errant `any`s.